requirements.txt
anyio==4.1.0
Brotli==1.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
future==0.18.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
iso8601==2.1.0
joblib==1.3.2
mutagen==1.47.0
nltk==3.8.1
PyAudio==0.2.14
pycryptodomex==3.19.0
pyserial==3.5
python-vlc==3.0.20123
pyttsx3==2.90
PyYAML==6.0.1
rake-nltk==1.0.6
regex==2023.10.3
requests==2.31.0
sniffio==1.3.0
SpeechRecognition==3.10.0
tqdm==4.66.1
urllib3==2.1.0
websockets==12.0
youtube-search-python==1.6.6
yt-dlp==2023.11.16
